runtime_bridge: lift sync→async helper to its own module, dedupe 3 inline call sites
`bridge_sync_to_async` lived as a file-local helper in `handle.rs`.
Three other call sites (`tombstones::cache`, `wal::pipeline`,
`superfile::vector::reader`) reimplemented the same `block_in_place +
Handle::block_on` / synthetic `current_thread` fallback inline, with
bespoke (and unreachable in practice) error mapping for the
runtime-build branch.

Moved the helper to a new crate-internal `runtime_bridge` module
with documentation covering the `multi_thread`-only requirement and
the rationale that motivates it (Python/asyncio consumers, rayon
workers, sync CLI callers). The four call sites now share one
implementation; the defensive runtime-build error mapping is gone —
the helper's `.expect("invariant: ...")` is the right call (tokio
runtime build only fails on OS resource exhaustion).

Net -67 lines. Writer-side sites at writer.rs:724/802/1310/1872
use a different shape (long-lived shared runtime via
`inner.sql_runtime()`) and are deliberately not collapsed in.
